+++ modules/core/src/main/scala-2.13/ghyll/CodecInstances.scala
@@ -1,0 +1,3 @@
+package ghyll
+
+trait CodecInstances