+++ modules/core/src/main/scala-2.13/ghyll/derivation/auto.scala
@@ -1,8 +1,11 @@
 package ghyll.derivation
 
 import ghyll._
+import shapeless.Lazy
 
 object auto {
-  implicit def autoDeriveDecoder[A](implicit d: DerivedDecoder[A]): Decoder[A] = deriveDecoder[A]
-  implicit def autoDeriveEncoder[A](implicit d: DerivedEncoder[A]): Encoder[A] = deriveEncoder[A]
+  implicit def autoDeriveDecoder[A: DerivedDecoder]: Decoder[A] = deriveDecoder[A]
+  implicit def autoDeriveEncoder[A: DerivedEncoder]: Encoder[A] = deriveEncoder[A]
+  implicit def autoDeriveCodec[A](implicit d: Lazy[DerivedDecoder[A]], e: Lazy[DerivedEncoder[A]]): Codec[A] =
+    deriveCodec[A](d.value, e.value)
 }