+++ modules/core/src/main/scala-3/ghyll/CodecInstances.scala
@@ -1,0 +1,7 @@
+package ghyll
+
+import ghyll.derivation.{DerivedDecoder, DerivedEncoder}
+
+trait CodecInstances {
+  inline given derived[A: DerivedDecoder: DerivedEncoder]: Codec[A] = Codec[A]
+}