+++ modules/core/src/main/scala/ghyll/derivation/Derivation.scala
@@ -1,9 +1,11 @@
 package ghyll.derivation
 
-import ghyll.{Decoder, Encoder}
+import ghyll.{Codec, Decoder, Encoder}
 
 private[ghyll] trait Derivation {
   def deriveDecoder[A](implicit d: DerivedDecoder[A]): Decoder[A] = d
 
   def deriveEncoder[A](implicit e: DerivedEncoder[A]): Encoder[A] = e
+
+  def deriveCodec[A: DerivedDecoder: DerivedEncoder]: Codec[A] = Codec[A]
 }