+++ modules/core/src/test/scala-3/ghyll/derivation/AutoDerivationSpec.scala
@@ -1,39 +1,45 @@
 package ghyll.derivation
 
 import cats.effect.IO
-import ghyll.Decoder
+import ghyll.{Codec, TestDecoder, TestEncoder}
 import ghyll.Utils.createReader
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
-class AutoDerivationSpec extends AnyWordSpec with Matchers {
-  // Todo change this so that Decoder can be derived directly
-  case class Foo(bar: String) derives DerivedDecoder
-  case class WrappedFoo(foo: Foo) derives DerivedDecoder
-
-  case class Response[A](data: A) derives DerivedDecoder
-
+class AutoDerivationSpec extends AnyWordSpec with Matchers with TestDecoder with TestEncoder {
   "deriving Decoder" should {
     "derive a valid decoder" when {
-      "a case class has only scalar attributes" in {
-        check(Foo("baz"), """{"bar": "baz"}""")
+      "derive decoders automatically for case classes" in {
+        case class Foo(bar: String) derives DerivedDecoder
+        case class WrappedFoo(foo: Foo) derives DerivedDecoder
+
+        case class Response[A](data: A) derives DerivedDecoder
+
+        testDecoder(Response(WrappedFoo(Foo("baz"))), """{"data": {"foo": {"bar": "baz"}}}""")
       }
 
-      "case classes are nested" in {
-        check(WrappedFoo(Foo("baz")), """{"foo": {"bar": "baz"}}""")
+      "derive encoders automatically for case classes" in {
+        case class Foo(bar: String) derives DerivedEncoder
+        case class WrappedFoo(foo: Foo) derives DerivedEncoder
+
+        case class Response[A](data: A) derives DerivedEncoder
+        testEncoder(Response(WrappedFoo(Foo("baz"))), """{"data": {"foo": {"bar": "baz"}}}""")
       }
 
-      "a case class has generic attributes" in {
-        check(Response(Foo("baz")), """{"data": {"bar": "baz"}}""")
+      "derive codecs automatically for case classes" in {
+        case class Foo(bar: String) derives Codec
+        case class WrappedFoo(foo: Foo) derives Codec
+
+        case class Response[A](data: A) derives Codec
+
+        def test[A: Codec](value: A, json: String) = {
+          testDecoder(value, json)
+          testEncoder(value, json)
+        }
+
+        test(Response(WrappedFoo(Foo("baz"))), """{"data": {"foo": {"bar": "baz"}}}""")
       }
+
     }
-  }
-
-  def check[T](expected: T, json: String)(implicit decoder: Decoder[T]) = {
-    createReader(json)
-      .use(reader => IO.delay(decoder.decode(reader)))
-      .map(_ should be(Right(expected)))
-      .unsafeRunSync()
-
   }
 }