+++ modules/core/src/test/scala/ghyll/derivation/DerivationSpec.scala
@@ -1,10 +1,10 @@
 package ghyll.derivation
 
-import ghyll.{Decoder, Encoder}
+import ghyll.{Codec, Decoder, Encoder, TestDecoder, TestEncoder}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
-class DerivationSpec extends AnyWordSpec with Matchers with Derivation {
+class DerivationSpec extends AnyWordSpec with Matchers with Derivation with TestDecoder with TestEncoder {
   case class Foo(bar: String)
   case class WrappedFoo(foo: Foo)
 
@@ -13,18 +13,24 @@ class DerivationSpec extends AnyWordSpec with Matchers with Derivation {
   "deriveDecoder" should {
     "derive a decoder" when {
       "a case class has only scalar attributes" in {
-        deriveDecoder[Foo] should be(a[Decoder[_]])
+        implicit val decoder: Decoder[Foo] = deriveDecoder[Foo]
+
+        testDecoder(Foo("baz"), """{"bar": "baz"}""")
       }
 
       "case classes are nested" in {
         implicit val fooDecoder: Decoder[Foo] = deriveDecoder[Foo]
-        deriveDecoder[WrappedFoo] should be(a[Decoder[_]])
+        implicit def responseDecoder[A: Decoder]: Decoder[Response[A]] = deriveDecoder
+
+        testDecoder(Response(Foo("baz")), """{"data": {"bar": "baz"}}""")
 
       }
 
       "a case class has generic attributes" in {
         implicit val fooDecoder: Decoder[Foo] = deriveDecoder[Foo]
-        deriveDecoder[Response[Foo]] should be(a[Decoder[_]])
+        implicit def responseDecoder[A: Decoder]: Decoder[Response[A]] = deriveDecoder
+
+        testDecoder(Response(Foo("baz")), """{"data": {"bar": "baz"}}""")
       }
     }
   }
@@ -32,18 +38,48 @@ class DerivationSpec extends AnyWordSpec with Matchers with Derivation {
   "deriveEncoder" should {
     "derive an encoder" when {
       "a case class has only scalar attributes" in {
-        deriveEncoder[Foo] should be(a[Encoder[_]])
+        implicit val fooEncoder: Encoder[Foo] = deriveEncoder[Foo]
+        testEncoder(Foo("baz"), """{"bar": "baz"}""")
       }
 
       "case classes are nested" in {
         implicit val fooEncoder: Encoder[Foo] = deriveEncoder[Foo]
-        deriveEncoder[WrappedFoo] should be(a[Encoder[_]])
+        implicit val wrappedFooEncoder: Encoder[WrappedFoo] = deriveEncoder[WrappedFoo]
 
+        testEncoder(WrappedFoo(Foo("baz")), """{"foo": {"bar": "baz"}}""")
       }
 
       "a case class has generic attributes" in {
         implicit val fooEncoder: Encoder[Foo] = deriveEncoder[Foo]
-        deriveEncoder[Response[Foo]] should be(a[Encoder[_]])
+        implicit val responseEncoder: Encoder[Response[Foo]] = deriveEncoder[Response[Foo]]
+
+        testEncoder(Response(Foo("baz")), """{"data": {"bar": "baz"}}""")
+      }
+    }
+  }
+
+  "deriveCodec" should {
+    "derive an codec" when {
+      "a case class has only scalar attributes" in {
+        implicit val fooCodec: Codec[Foo] = deriveCodec[Foo]
+        testDecoder(Foo("baz"), """{"bar": "baz"}""")
+        testEncoder(Foo("baz"), """{"bar": "baz"}""")
+      }
+
+      "case classes are nested" in {
+        implicit val fooCodec: Codec[Foo] = deriveCodec[Foo]
+        implicit val wrappedFooCodec: Codec[WrappedFoo] = deriveCodec[WrappedFoo]
+
+        testDecoder(WrappedFoo(Foo("baz")), """{"foo": {"bar": "baz"}}""")
+        testEncoder(WrappedFoo(Foo("baz")), """{"foo": {"bar": "baz"}}""")
+      }
+
+      "a case class has generic attributes" in {
+        implicit val fooCodec: Codec[Foo] = deriveCodec[Foo]
+        implicit val responseCodec: Codec[Response[Foo]] = deriveCodec[Response[Foo]]
+
+        testDecoder(Response(Foo("baz")), """{"data": {"bar": "baz"}}""")
+        testEncoder(Response(Foo("baz")), """{"data": {"bar": "baz"}}""")
       }
     }
   }