+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - it is now possible to decode an object under a given JSON path
 - `Decoder` can be fully automatically derived
 - decode* shortcuts that can decode from `java.io.File`
+- `Encoder` capabilities that match the `Decoder`'s functionality
+  (auto derivation, etc)
 
 ### Fixed